Removed code duplication, moved wkt to separate dir
