+++ wkt/__init__.py
@@ -1,0 +1,1 @@
+"""WKT wrapper for Shapely for working with WKT type objects"""
